+++ demisto_sdk/commands/common/constants.py
@@ -1606,3 +1606,4 @@ class ParameterType(Enum):
 
 
 NO_TESTS_DEPRECATED = 'No tests (deprecated)'
+NATIVE_IMAGE_FILE_NAME = 'docker_native_image_config.json'