+++ demisto_sdk/commands/content_graph/objects/integration.py
@@ -1,3 +1,4 @@
+import logging
 from typing import TYPE_CHECKING, List
 
 from demisto_sdk.commands.content_graph.objects.base_content import BaseContent
@@ -8,8 +9,11 @@ if TYPE_CHECKING:
 
 from pydantic import Field
 
+from demisto_sdk.commands.common.constants import MarketplaceVersions
 from demisto_sdk.commands.content_graph.common import ContentType, RelationshipType
 from demisto_sdk.commands.content_graph.objects.integration_script import IntegrationScript
+
+logger = logging.getLogger('demisto-sdk')
 
 
 class Command(BaseContent, content_type=ContentType.COMMAND):  # type: ignore[call-arg]
@@ -70,3 +74,17 @@ class Integration(IntegrationScript, content_type=ContentType.INTEGRATION):  # t
             "category": True,
             "commands": {"name": True, "description": True},
         }
+
+    def prepare_for_upload(self, marketplace: MarketplaceVersions = MarketplaceVersions.XSOAR, **kwargs) -> dict:
+        data = super().prepare_for_upload(marketplace, **kwargs)
+
+        if supported_native_images := self.get_supported_native_images(
+            marketplace=marketplace,
+            ignore_native_image=kwargs.get('ignore_native_image') or False
+        ):
+            logger.debug(
+                f'Adding the following native images {supported_native_images} to integration {self.object_id}'
+            )
+            data['script']['nativeImage'] = supported_native_images
+
+        return data